The Great Rewrite (fixes #36, #61, #94, #101)
Rewrite of the file model and pulling mechanism. Needs lots of cleanup
and bugfixes, now...

+++ protocol/message_types.go
@@ -9,7 +9,7 @@ type FileInfo struct {
 	Name     string // max:1024
 	Flags    uint32
 	Modified int64
-	Version  uint32
+	Version  uint64
 	Blocks   []BlockInfo // max:100000
 }
